+++ mozilla/mobile/android/modules/Sanitizer.jsm
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-/*globals LoadContextInfo, FormHistory, Accounts */
+/* globals LoadContextInfo, FormHistory, Accounts */
 
 var Cc = Components.classes;
 var Ci = Components.interfaces;
@@ -13,7 +13,6 @@ Cu.import("resource://gre/modules/Services.jsm");
 Cu.import("resource://gre/modules/XPCOMUtils.jsm");
 Cu.import("resource://gre/modules/LoadContextInfo.jsm");
 Cu.import("resource://gre/modules/FormHistory.jsm");
-Cu.import("resource://gre/modules/Messaging.jsm");
 Cu.import("resource://gre/modules/Task.jsm");
 Cu.import("resource://gre/modules/Downloads.jsm");
 Cu.import("resource://gre/modules/osfile.jsm");
@@ -21,6 +20,10 @@ Cu.import("resource://gre/modules/Accounts.jsm");
 
 XPCOMUtils.defineLazyModuleGetter(this, "DownloadIntegration",
                                   "resource://gre/modules/DownloadIntegration.jsm");
+XPCOMUtils.defineLazyModuleGetter(this, "EventDispatcher",
+                                  "resource://gre/modules/Messaging.jsm");
+XPCOMUtils.defineLazyModuleGetter(this, "TelemetryStopwatch",
+                                  "resource://gre/modules/TelemetryStopwatch.jsm");
 
 function dump(a) {
   Services.console.logStringMessage(a);
@@ -30,63 +33,89 @@ this.EXPORTED_SYMBOLS = ["Sanitizer"];
 
 function Sanitizer() {}
 Sanitizer.prototype = {
-  clearItem: function (aItemName)
-  {
+  clearItem: function(aItemName, startTime) {
+    // Only a subset of items support deletion with startTime.
+    // Those who do not will be rejected with error message.
+    if (typeof startTime != "undefined") {
+      switch (aItemName) {
+        // Normal call to DownloadFiles remove actual data from storage, but our web-extension consumer
+        // deletes only download history. So, for this reason we are passing a flag 'deleteFiles'.
+        case "downloadHistory":
+          return this._clear("downloadFiles", { startTime, deleteFiles: false });
+        case "formdata":
+          return this._clear(aItemName, { startTime });
+        default:
+          return Promise.reject({message: `Invalid argument: ${aItemName} does not support startTime argument.`});
+      }
+    } else {
+      return this._clear(aItemName);
+    }
+  },
+
+ _clear: function(aItemName, options) {
     let item = this.items[aItemName];
     let canClear = item.canClear;
     if (typeof canClear == "function") {
       canClear(function clearCallback(aCanClear) {
         if (aCanClear)
-          item.clear();
+          return item.clear(options);
       });
     } else if (canClear) {
-      item.clear();
+      return item.clear(options);
     }
   },
 
   items: {
     cache: {
-      clear: function ()
-      {
+      clear: function() {
         return new Promise(function(resolve, reject) {
+          let refObj = {};
+          TelemetryStopwatch.start("FX_SANITIZE_CACHE", refObj);
+
           var cache = Cc["@mozilla.org/netwerk/cache-storage-service;1"].getService(Ci.nsICacheStorageService);
           try {
             cache.clear();
-          } catch(er) {}
+          } catch (er) {}
 
           let imageCache = Cc["@mozilla.org/image/tools;1"].getService(Ci.imgITools)
                                                            .getImgCacheForDocument(null);
           try {
             imageCache.clearCache(false); // true=chrome, false=content
-          } catch(er) {}
+          } catch (er) {}
 
+          TelemetryStopwatch.finish("FX_SANITIZE_CACHE", refObj);
           resolve();
         });
       },
 
-      get canClear()
-      {
+      get canClear() {
         return true;
       }
     },
 
     cookies: {
-      clear: function ()
-      {
+      clear: function() {
         return new Promise(function(resolve, reject) {
+          let refObj = {};
+          TelemetryStopwatch.start("FX_SANITIZE_COOKIES_2", refObj);
+
           Services.cookies.removeAll();
+
+          TelemetryStopwatch.finish("FX_SANITIZE_COOKIES_2", refObj);
           resolve();
         });
       },
 
-      get canClear()
-      {
+      get canClear() {
         return true;
       }
     },
 
     siteSettings: {
       clear: Task.async(function* () {
+        let refObj = {};
+        TelemetryStopwatch.start("FX_SANITIZE_SITESETTINGS", refObj);
+
         // Clear site-specific permissions like "Allow this site to open popups"
         Services.perms.removeAll();
 
@@ -113,44 +142,44 @@ Sanitizer.prototype = {
             }
           });
         });
+        TelemetryStopwatch.finish("FX_SANITIZE_SITESETTINGS", refObj);
       }),
 
-      get canClear()
-      {
+      get canClear() {
         return true;
       }
     },
 
     offlineApps: {
-      clear: function ()
-      {
+      clear: function() {
         return new Promise(function(resolve, reject) {
           var cacheService = Cc["@mozilla.org/netwerk/cache-storage-service;1"].getService(Ci.nsICacheStorageService);
           var appCacheStorage = cacheService.appCacheStorage(LoadContextInfo.default, null);
           try {
             appCacheStorage.asyncEvictStorage(null);
-          } catch(er) {}
+          } catch (er) {}
 
           resolve();
         });
       },
 
-      get canClear()
-      {
+      get canClear() {
           return true;
       }
     },
 
     history: {
-      clear: function ()
-      {
-        return Messaging.sendRequestForResult({ type: "Sanitize:ClearHistory" })
+      clear: function() {
+        let refObj = {};
+        TelemetryStopwatch.start("FX_SANITIZE_HISTORY", refObj);
+
+        return EventDispatcher.instance.sendRequestForResult({ type: "Sanitize:ClearHistory" })
           .catch(e => Cu.reportError("Java-side history clearing failed: " + e))
           .then(function() {
+            TelemetryStopwatch.finish("FX_SANITIZE_HISTORY", refObj);
             try {
-              Services.obs.notifyObservers(null, "browser:purge-session-history", "");
-            }
-            catch (e) { }
+              Services.obs.notifyObservers(null, "browser:purge-session-history");
+            } catch (e) { }
 
             try {
               var predictor = Cc["@mozilla.org/network/predictor;1"].getService(Ci.nsINetworkPredictor);
@@ -159,38 +188,64 @@ Sanitizer.prototype = {
           });
       },
 
-      get canClear()
-      {
+      get canClear() {
         // bug 347231: Always allow clearing history due to dependencies on
         // the browser:purge-session-history notification. (like error console)
         return true;
       }
     },
 
-    searchHistory: {
-      clear: function ()
-      {
-        return Messaging.sendRequestForResult({ type: "Sanitize:ClearHistory", clearSearchHistory: true })
-          .catch(e => Cu.reportError("Java-side search history clearing failed: " + e))
+    openTabs: {
+      clear: function() {
+        let refObj = {};
+        TelemetryStopwatch.start("FX_SANITIZE_OPENWINDOWS", refObj);
+
+        return EventDispatcher.instance.sendRequestForResult({ type: "Sanitize:OpenTabs" })
+          .catch(e => Cu.reportError("Java-side tab clearing failed: " + e))
+          .then(function() {
+            try {
+              // clear "Recently Closed" tabs in Android App
+              Services.obs.notifyObservers(null, "browser:purge-session-tabs");
+            } catch (e) { }
+            TelemetryStopwatch.finish("FX_SANITIZE_OPENWINDOWS", refObj);
+          });
       },
 
-      get canClear()
-      {
+      get canClear() {
+        return true;
+      }
+    },
+
+    searchHistory: {
+      clear: function() {
+        return EventDispatcher.instance.sendRequestForResult({ type: "Sanitize:ClearHistory", clearSearchHistory: true })
+          .catch(e => Cu.reportError("Java-side search history clearing failed: " + e));
+      },
+
+      get canClear() {
         return true;
       }
     },
 
     formdata: {
-      clear: function ()
-      {
+      clear: function({ startTime = 0 } = {}) {
         return new Promise(function(resolve, reject) {
-          FormHistory.update({ op: "remove" });
+          let refObj = {};
+          TelemetryStopwatch.start("FX_SANITIZE_FORMDATA", refObj);
+
+          // Conver time to microseconds
+          let time = startTime * 1000;
+          FormHistory.update({
+            op: "remove",
+            firstUsedStart: time
+          });
+
+          TelemetryStopwatch.finish("FX_SANITIZE_FORMDATA", refObj);
           resolve();
         });
       },
 
-      canClear: function (aCallback)
-      {
+      canClear: function(aCallback) {
         let count = 0;
         let countDone = {
           handleResult: function(aResult) { count = aResult; },
@@ -202,7 +257,10 @@ Sanitizer.prototype = {
     },
 
     downloadFiles: {
-      clear: Task.async(function* () {
+      clear: Task.async(function* ({ startTime = 0, deleteFiles = true} = {}) {
+        let refObj = {};
+        TelemetryStopwatch.start("FX_SANITIZE_DOWNLOADS", refObj);
+
         let list = yield Downloads.getList(Downloads.ALL);
         let downloads = yield list.getAll();
         var finalizePromises = [];
@@ -213,8 +271,10 @@ Sanitizer.prototype = {
         for (let download of downloads) {
           // Remove downloads that have been canceled, even if the cancellation
           // operation hasn't completed yet so we don't check "stopped" here.
-          // Failed downloads with partial data are also removed.
-          if (download.stopped && (!download.hasPartialData || download.error)) {
+          // Failed downloads with partial data are also removed. The startTime
+          // check is provided for addons that may want to delete only recent downloads.
+          if (download.stopped && (!download.hasPartialData || download.error) &&
+              download.startTime.getTime() >= startTime) {
             // Remove the download first, so that the views don't get the change
             // notifications that may occur during finalization.
             yield list.remove(download);
@@ -224,74 +284,74 @@ Sanitizer.prototype = {
             // processing the other downloads in the list.
             finalizePromises.push(download.finalize(true).then(() => null, Cu.reportError));
 
-            // Delete the downloaded files themselves.
-            OS.File.remove(download.target.path).then(() => null, ex => {
-              if (!(ex instanceof OS.File.Error && ex.becauseNoSuchFile)) {
-                Cu.reportError(ex);
-              }
-            });
+            if (deleteFiles) {
+              // Delete the downloaded files themselves.
+              OS.File.remove(download.target.path).then(() => null, ex => {
+                if (!(ex instanceof OS.File.Error && ex.becauseNoSuchFile)) {
+                  Cu.reportError(ex);
+                }
+              });
+            }
           }
         }
 
         yield Promise.all(finalizePromises);
         yield DownloadIntegration.forceSave();
+        TelemetryStopwatch.finish("FX_SANITIZE_DOWNLOADS", refObj);
       }),
 
-      get canClear()
-      {
+      get canClear() {
         return true;
       }
     },
 
     passwords: {
-      clear: function ()
-      {
+      clear: function() {
         return new Promise(function(resolve, reject) {
           Services.logins.removeAllLogins();
           resolve();
         });
       },
 
-      get canClear()
-      {
+      get canClear() {
         let count = Services.logins.countLogins("", "", ""); // count all logins
         return (count > 0);
       }
     },
 
     sessions: {
-      clear: function ()
-      {
+      clear: function() {
         return new Promise(function(resolve, reject) {
+          let refObj = {};
+          TelemetryStopwatch.start("FX_SANITIZE_SESSIONS", refObj);
+
           // clear all auth tokens
           var sdr = Cc["@mozilla.org/security/sdr;1"].getService(Ci.nsISecretDecoderRing);
           sdr.logoutAndTeardown();
 
           // clear FTP and plain HTTP auth sessions
-          Services.obs.notifyObservers(null, "net:clear-active-logins", null);
+          Services.obs.notifyObservers(null, "net:clear-active-logins");
 
+          TelemetryStopwatch.finish("FX_SANITIZE_SESSIONS", refObj);
           resolve();
         });
       },
 
-      get canClear()
-      {
+      get canClear() {
         return true;
       }
     },
 
     syncedTabs: {
-      clear: function ()
-      {
-        return Messaging.sendRequestForResult({ type: "Sanitize:ClearSyncedTabs" })
+      clear: function() {
+        return EventDispatcher.instance.sendRequestForResult({ type: "Sanitize:ClearSyncedTabs" })
           .catch(e => Cu.reportError("Java-side synced tabs clearing failed: " + e));
       },
 
-      canClear: function(aCallback)
-      {
+      canClear: function(aCallback) {
         Accounts.anySyncAccountsExist().then(aCallback)
           .catch(function(err) {
-            Cu.reportError("Java-side synced tabs clearing failed: " + err)
+            Cu.reportError("Java-side synced tabs clearing failed: " + err);
             aCallback(false);
           });
       }
